third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/no-attribute-cross-origin-bc-containers.https-expected.txt

This is a testharness.js-based test.
FAIL iframe assert_array_equals: property 0, expected true but got false
FAIL frame assert_array_equals: property 0, expected true but got false
FAIL object assert_array_equals: property 0, expected true but got false
FAIL embed assert_array_equals: property 0, expected true but got false
Harness: the test ran to completion.
